whatever

+++ ServoBrainV0_1.X/AccelTable.h
@@ -19,19 +19,18 @@ typedef enum { Positive, Negative } AccelDirection;
 
 typedef struct AccelTable
 {
-    int32_t Velocity[QTY_ENTRIES];
-    int32_t Negative[QTY_ENTRIES];
-    int32_t Positive[QTY_ENTRIES];
+    int16_t Velocity[QTY_ENTRIES];
+    int16_t Negative[QTY_ENTRIES];
+    int16_t Positive[QTY_ENTRIES];
 } AccelTable;
-
 
 void AccelTableBuildDefault();
 void DebugAccelTable();
-int32_t GetPositiveAccel(const State *state);
-int32_t GetNegativeAccel(const State *state);
+int16_t GetPositiveAccel(const State *state);
+int16_t GetNegativeAccel(const State *state);
 
-extern int32_t (*accelFuncPositive)(const State* state);
-extern int32_t (*accelFuncNegative)(const State* state);
+extern int16_t (*accelFuncPositive)(const State* state);
+extern int16_t (*accelFuncNegative)(const State* state);
